Restaurant PoS System
This is a basic Restaurant Point of Sale system that I created using only Java. I made this so I could improve on my programming skills and learn more about java while doing so.

This project makes use of the builit in swing class to build a GUI that the user can use with ease.

I hope you like it.